+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.5.3 (SNAPSHOT)
 * Modify jscoverage-based coverage target to always reinstrument, to allow multiple coverage target calls without clean
+* Add alias target support
 
 ### 0.5.2
 * Add --verbose opt to display make arguments and shell commands